Cleanup. There's still a lot of memory leaks to fix.

+++ user/fib/src/fib.c
@@ -13,8 +13,8 @@ int main(const int argc, char **argv)
 {
     putchar('\n');
     printf("Calculating fibonacci of 42 using recursion.\n");
-    const uint64_t result = fib(42);
-    printf("\nResult: %llu", result);
+    const uint32_t result = fib(42);
+    printf("\nResult: %lu", result);
 
     // uint64_t a = 0, b = 1, c, i, n = 100;
     // for (i = 0; i < n; i++) {
